Speedup of FF, using sparse action
Significant speedup for translating standard tower to simple extension.
This uses that the primitve element (a monomial in the tower basis)
has a very sparse action on the tower basis.
The sparse representation of this action allows for a fast computation
of the minimal polynomial (via Berlekamp-Massey) and of a matrix whose
rows represent the powers of the primitive element on the tower basis.