Find standard polynomials with NTL.
A version of `SteinitzNumberForPrimeDegree` that uses NTL to find
standard polynomials in characteristic 2, resp. p < 2^50 for prime
degree r over the prime field GF(p).

The external programs in the ntl/ directory need to be compiled using
libntl (see head of .cc files).

The NTL version of Alg. 5.5 seems in particular faster for
   p = 2
   large degree (> 500) for p < 256
   p > 256

Still experimental. Not documented at this stage.